Toggle extra field in profile

+++ include/functions.php
@@ -51,13 +51,14 @@ function settings_address()
 
 	$arr_settings = array();
 
-	if(IS_SUPER_ADMIN)
+	if(IS_SUPER_ADMIN && is_multisite())
 	{
-		$arr_settings['setting_address_site_wide'] = __("Use the master table on all sites", 'lang_address');
+		$arr_settings['setting_address_site_wide'] = __("Use Master Table on All Sites", 'lang_address');
 	}
 
-	$arr_settings['setting_address_extra'] = __("Name for extra address field", 'lang_address');
-	$arr_settings['setting_show_member_id'] = __("Show memberID", 'lang_address');
+	$arr_settings['setting_address_extra'] = __("Name for Extra Address Field", 'lang_address');
+	$arr_settings['setting_address_extra_profile'] = __("Display Settings for Extra in Profile", 'lang_address');
+	$arr_settings['setting_show_member_id'] = __("Display Member ID", 'lang_address');
 
 	show_settings_fields(array('area' => $options_area, 'settings' => $arr_settings));
 }
@@ -94,6 +95,14 @@ function setting_address_extra_callback()
 	echo show_textfield(array('name' => $setting_key, 'value' => $option));
 }
 
+function setting_address_extra_profile_callback()
+{
+	$setting_key = get_setting_key(__FUNCTION__);
+	$option = get_option($setting_key, 'yes');
+
+	echo show_select(array('data' => get_yes_no_for_select(), 'name' => $setting_key, 'value' => $option));
+}
+
 function menu_address()
 {
 	$menu_root = 'mf_address/';
@@ -119,58 +128,56 @@ function show_profile_address($user)
 {
 	global $wpdb;
 
-	$meta_address_permission = get_user_meta($user->ID, 'meta_address_permission', true);
-
-	$result = $wpdb->get_results("SELECT addressExtra FROM ".get_address_table_prefix()."address WHERE addressExtra != '' GROUP BY addressExtra");
-
-	if($wpdb->num_rows > 0)
+	if(IS_ADMIN && get_option_or_default('setting_address_extra_profile', 'yes') == 'yes')
 	{
-		echo "<table class='form-table'>
-			<tr class='user-address-permission-wrap'>
-				<th><label for='meta_address_permission'>".__("Address Permissions to Users", 'lang_address').":</label></th>
-				<td>";
+		$meta_address_permission = get_user_meta($user->ID, 'meta_address_permission', true);
 
-					$meta_address_permission = get_user_meta($user->ID, 'meta_address_permission', true);
-					$meta_address_permission = explode(",", $meta_address_permission);
+		$result = $wpdb->get_results("SELECT addressExtra FROM ".get_address_table_prefix()."address WHERE addressExtra != '' GROUP BY addressExtra");
 
-					$arr_data = array();
+		if($wpdb->num_rows > 0)
+		{
+			echo "<table class='form-table'>
+				<tr class='user-address-permission-wrap'>
+					<th><label for='meta_address_permission'>".__("Address Permissions to Users", 'lang_address').":</label></th>
+					<td>";
 
-					foreach($result as $r)
-					{
-						$strTableValue = $r->addressExtra;
+						$meta_address_permission = get_user_meta($user->ID, 'meta_address_permission', true);
+						$meta_address_permission = explode(",", $meta_address_permission);
 
-						if($strTableValue != '')
+						$arr_data = array();
+
+						foreach($result as $r)
 						{
-							$arr_data[$strTableValue] = $strTableValue;
-						}
-					}
+							$strTableValue = $r->addressExtra;
 
-					echo show_select(array('data' => $arr_data, 'name' => 'meta_address_permission[]', 'value' => $meta_address_permission))
-				."</td>
-			</tr>
-		</table>";
+							if($strTableValue != '')
+							{
+								$arr_data[$strTableValue] = $strTableValue;
+							}
+						}
+
+						echo show_select(array('data' => $arr_data, 'name' => 'meta_address_permission[]', 'value' => $meta_address_permission))
+					."</td>
+				</tr>
+			</table>";
+		}
 	}
 }
 
 function save_portfolio_address($user_id)
 {
-	if(current_user_can('update_core', $user_id))
+	if(IS_ADMIN && get_option_or_default('setting_address_extra_profile', 'yes') == 'yes')
 	{
-		save_register_address($user_id);
-	}
-}
+		$meta_address_permission = isset($_POST['meta_address_permission']) ? $_POST['meta_address_permission'] : "";
 
-function save_register_address($user_id)
-{
-	$meta_address_permission = isset($_POST['meta_address_permission']) ? $_POST['meta_address_permission'] : "";
+		if(is_array($meta_address_permission))
+		{
+			update_user_meta($user_id, 'meta_address_permission', implode(",", $meta_address_permission));
+		}
 
-	if(is_array($meta_address_permission))
-	{
-		update_user_meta($user_id, 'meta_address_permission', implode(",", $meta_address_permission));
-	}
-
-	else
-	{
-		delete_user_meta($user_id, 'meta_address_permission');
+		else
+		{
+			delete_user_meta($user_id, 'meta_address_permission');
+		}
 	}
 }
